Update npm deps, use code-splitting for Wasmer (#28)
This update reduces the size of resulting `dist/main.js` from 1.9 to 1.2 megabytes, which allows our main page to load faster. The rest of the code, which is coming from `@wasmer/wasi` and `@wasmer/wasmfs` is loaded asynchronously only when a user tries to compile any Swift in the editor view.